Remove package layer
