+++ metric2/meter/syncfloat64/syncfloat64.go
@@ -1,4 +1,4 @@
-package syncfloat64metric
+package syncfloat64
 
 import (
 	"context"
