+++ metric2/meter/syncint64/syncint64.go
@@ -1,4 +1,4 @@
-package asyncint64metric
+package syncint64
 
 import (
 	"context"
@@ -6,6 +6,8 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	metric "go.opentelemetry.io/otel/metric2"
 )
+
+// TODO instrument options
 
 type Meter struct {
 }
@@ -16,20 +18,19 @@ type Counter struct {
 type UpDownCounter struct {
 }
 
-type Gauge struct {
+type Histogram struct {
 }
 
 type Instrument interface {
 	metric.Instrument
 
-	Observe(ctx context.Context, x int64, attrs ...attribute.KeyValue)
 	Measure(x int64) metric.Measurement
 }
 
 var (
 	_ Instrument = Counter{}
 	_ Instrument = UpDownCounter{}
-	_ Instrument = Gauge{}
+	_ Instrument = Histogram{}
 )
 
 func (m Meter) Counter(name string) (Counter, error) {
@@ -40,17 +41,17 @@ func (m Meter) UpDownCounter(name string) (UpDownCounter, error) {
 	return UpDownCounter{}, nil
 }
 
-func (m Meter) Gauge(name string) (Gauge, error) {
-	return Gauge{}, nil
+func (m Meter) Histogram(name string) (Histogram, error) {
+	return Histogram{}, nil
 }
 
-func (c Counter) Observe(ctx context.Context, x int64, attrs ...attribute.KeyValue) {
+func (c Counter) Add(ctx context.Context, x int64, attrs ...attribute.KeyValue) {
 }
 
-func (u UpDownCounter) Observe(ctx context.Context, x int64, attrs ...attribute.KeyValue) {
+func (u UpDownCounter) Add(ctx context.Context, x int64, attrs ...attribute.KeyValue) {
 }
 
-func (g Gauge) Observe(ctx context.Context, x int64, attrs ...attribute.KeyValue) {
+func (h Histogram) Record(ctx context.Context, x int64, attrs ...attribute.KeyValue) {
 }
 
 func (c Counter) Measure(x int64) metric.Measurement {
@@ -61,6 +62,6 @@ func (u UpDownCounter) Measure(x int64) metric.Measurement {
 	return metric.Measurement{}
 }
 
-func (g Gauge) Measure(x int64) metric.Measurement {
+func (h Histogram) Measure(x int64) metric.Measurement {
 	return metric.Measurement{}
 }
